+++ library/source/Quaternion.cpp
@@ -4,7 +4,7 @@
 
 Quaternion::Quaternion() = default;
 
-Quaternion::Quaternion(float scalar, const Vector3 &vector) {
+Quaternion::Quaternion(double scalar, const Vector3 &vector) {
     this->scalar = scalar;
     this->vector = vector;
 }
@@ -29,12 +29,12 @@ Quaternion Quaternion::operator-(const Quaternion &quaternion) const {
 }
 
 Quaternion Quaternion::operator*(const Quaternion &quaternion) const {
-    float s = scalar * quaternion.scalar - vector.dotProduct(quaternion.vector);
+    double s = scalar * quaternion.scalar - vector.dotProduct(quaternion.vector);
     Vector3 v = vector * quaternion.scalar + quaternion.vector * scalar + vector.crossProduct(quaternion.vector);
     return {s, v};
 }
 
-Quaternion Quaternion::operator*(float f) const {
+Quaternion Quaternion::operator*(double f) const {
     return {scalar * f, vector * f};
 }
 
@@ -55,26 +55,26 @@ Quaternion Quaternion::operator*=(const Quaternion &quaternion) {
     return *this;
 }
 
-Quaternion Quaternion::operator*=(float f) {
+Quaternion Quaternion::operator*=(double f) {
     *this = *this * f;
     return *this;
 }
 
-float Quaternion::norm() const {
-    float s = scalar * scalar;
-    float v = vector.x * vector.x + vector.y * vector.y + vector.z * vector.z;
+double Quaternion::norm() const {
+    double s = scalar * scalar;
+    double v = vector.x * vector.x + vector.y * vector.y + vector.z * vector.z;
     return std::sqrt(s + v);
 }
 
 Quaternion Quaternion::normalize() {
-    float norm = this->norm();
+    double norm = this->norm();
     if(norm == 0) return *this;
     *this *= 1/norm;
     return *this;
 }
 
 Quaternion Quaternion::normalized() const {
-    float norm = this->norm();
+    double norm = this->norm();
     if(norm == 0) return *this;
     return *this * (1/norm);
 }
@@ -84,7 +84,7 @@ Quaternion Quaternion::conjugate() const {
 }
 
 Quaternion Quaternion::inverse() const {
-    float absoluteValue = this->norm();
+    double absoluteValue = this->norm();
 
     if(absoluteValue == 0) return *this;
 
@@ -97,10 +97,10 @@ Quaternion Quaternion::inverse() const {
 }
 
 Quaternion Quaternion::convertToUnitNorm() {
-    float angle = scalar;
+    double angle = scalar;
     vector.normalize();
-    float s = std::cos(angle * 0.5f);
-    Vector3 v = vector * std::sin(angle * 0.5f);
+    double s = std::cos(angle * 0.5);
+    Vector3 v = vector * std::sin(angle * 0.5);
     return {s, v};
 }
